Replace hard-copy of phpmailer with composer
1. Why is this change necessary?
A phpmailer vulnerability for version < 5.2.18 has been reported
in issues #2 and so we need to update the library version as per
vendor recommendations.

2. How does it address the issue?
By ditching the hard-copy version of the library and using an
updated version from composer. Thought is this will ease updating
coming to future versions.

Relates to #2