Minify HTML templates
1. Why is this change necessary?
There's a limit with the number of characters accepted per line in
an email as seen in https://code.djangoproject.com/ticket/22561.

PHPMailer claims it can wrap the lines and such, but it fails in an
unpredictable way and tries to switch to 'quoted-printable' encoding
when appropriate.

There's work going in the v6 branch of PHPMailer to tackle this. But
meanwhile minifying the HTML and throwing it to the library seems to
circumvent the issue.

As such this is just a temporary fix so we tell composer to require
from this branch, meanwhile we await official release